Move APICall to models
